# Fixing logout -which also fixes mod_logged- (Thank you Dennis!)
git-svn-id: http://joomlacode.org/svn/joomla/development/trunk@16271 6f6e1ebd-4c2b-0410-823f-f34bde69bce9

+++ plugins/user/joomla/joomla.php
@@ -117,17 +117,21 @@ class plgUserJoomla extends JPlugin
 	 */
 	function onLogoutUser($user, $options = array())
 	{
-		$my = JFactory::getUser();
+		$my 		= &JFactory::getUser();
+		$session 	= &JFactory::getSession();
+		$app 		= &JFactory::getApplication();
+		
 		// Make sure we're a valid user first
-		if ($user['id'] == 0 && !$my->get('tmp_user')) return true;
-
+		if ($user['id'] == 0 && !$my->get('tmp_user')) {
+			return true;
+		}
+		
 		// Check to see if we're deleting the current session
-		if ($my->get('id') == $user['id']) {
+		if ($my->get('id') == $user['id'] && $options['clientid'] == $app->getClientId()) {
 			// Hit the user last visit field
 			$my->setLastVisit();
 
 			// Destroy the php session for this user
-			$session = JFactory::getSession();
 			$session->destroy();
 		} else {
 			// Force logout all users with that userid
